3.C++-Project_Movie-Rental-Store
First, I created 3 h files and named each of them (customer, movie and movie rental). I created my customer class in the customer h file, my movie class in the movie h file, and my movie rental store class and properties in my movie rental h file. I created my objects in my movie_main.cpp file and printed them on the command line. I tried to fill in the movie number and movie list properties of my movie rental store class. To do this, I had to make all properties in the Movie rental store class static. Because I was going to change the values from the movie rental function. When I first ran it, I got an undefine definition error. After solving this error, I was always getting the following error (multiple definition). I found out why this happened. When writing the first value of static variables, we had to write it to the cpp file, not the h file. The reason for this is that when we call the h file 4 or 5 times, we also call the first value of the static variable 4-5 times. That's why I defined my static variables in the movie_lib.cpp file.